[issue-1357] remove procstats package
Types and functions from procstats package are moved to
internal/common package
